+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - bump `wechat_assets_picker` to 9.1.0
 - fixes the deprecated `ColorScheme.background` warnings
+- fixes empty gallery on restorable picker first open [#45](https://github.com/LeGoffMael/insta_assets_picker/pull/45)
 
 ## 2.3.0